Implemented postItem method in REST controllers to INSERT and UPDATE entities with relationships in a single REST request, using Java Reflection API and JPA. Implemented custom getItem method to return different results for different types of entities.

+++ src/main/kotlin/ru/itport/yourspendings/controller/ProductCategoriesController.kt
@@ -36,9 +36,9 @@ class ProductCategoriesController:EntityController<ProductCategory>("ProductCate
         return result.subList(req.skip!!, if (req.skip!!+req.limit!!>result.size) result.size; else req.skip!!+req.limit!!)
     }
 
-    private fun getParentList(item:ProductCategory):ArrayList<Long> {
+    private fun getParentList(item:ProductCategory):ArrayList<Int> {
         var parent = item.parent
-        var result = ArrayList<Long>()
+        var result = ArrayList<Int>()
         if (parent == null) return result;
         while (parent != null) {
             result.add(parent.uid!!)
